+++ miso-service/src/main/java/uk/ac/bbsrc/tgac/miso/service/impl/DefaultLibraryService.java
@@ -22,6 +22,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.transaction.support.TransactionTemplate;
 
 import com.eaglegenomics.simlims.core.Note;
 import com.eaglegenomics.simlims.core.User;
@@ -73,12 +74,10 @@ import uk.ac.bbsrc.tgac.miso.core.util.PaginationFilter;
 import uk.ac.bbsrc.tgac.miso.core.util.Pluralizer;
 import uk.ac.bbsrc.tgac.miso.persistence.LibraryStore;
 import uk.ac.bbsrc.tgac.miso.persistence.SampleStore;
-import uk.ac.bbsrc.tgac.miso.persistence.impl.util.HibernateSessionManager;
-import uk.ac.bbsrc.tgac.miso.service.HibernateBulkSaveService;
 
 @Transactional(rollbackFor = Exception.class)
 @Service
-public class DefaultLibraryService implements HibernateBulkSaveService<Library>, LibraryService, PaginatedDataSource<Library> {
+public class DefaultLibraryService implements LibraryService, PaginatedDataSource<Library> {
 
   @Autowired
   private LibraryStore libraryDao;
@@ -125,7 +124,7 @@ public class DefaultLibraryService implements HibernateBulkSaveService<Library>,
   @Autowired
   private SopService sopService;
   @Autowired
-  private HibernateSessionManager hibernateSessionManager;
+  private TransactionTemplate transactionTemplate;
   @Value("${miso.autoGenerateIdentificationBarcodes}")
   private Boolean autoGenerateIdBarcodes;
 
@@ -769,8 +768,8 @@ public class DefaultLibraryService implements HibernateBulkSaveService<Library>,
   }
 
   @Override
-  public HibernateSessionManager getHibernateSessionManager() {
-    return hibernateSessionManager;
+  public TransactionTemplate getTransactionTemplate() {
+    return transactionTemplate;
   }
 
 }